SOLID-ize printer pt. 1
